스케줄러 수정
